requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
pandas==1.5.3
python-dotenv==1.0.0